Never use async generator outside a context manager.
