+++ byn/utils.py
@@ -126,3 +126,14 @@ async def alist(coro) -> list:
 async def atuple(coro) -> tuple:
     return tuple(await alist(coro))
 
+
+_not_set = object()
+
+async def anext(async_iter, default=_not_set):
+    try:
+        return await (await async_iter.__aiter__()).__anext__()
+    except StopAsyncIteration as e:
+        if default is _not_set:
+            raise e
+
+        return default
